Move major funcionalities to ScreenshotActivity
