+++ app/src/main/res/layout/activity_main.xml
@@ -27,11 +27,6 @@
                 android:layout_height="wrap_content"
                 android:text="disconnect" />
 
-            <ImageView
-                android:id="@+id/imageViewScreenshot"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
         </LinearLayout>
 
     </ScrollView>
